feature: refactor code, move user pilot functions to separate class.
